design(variables.scss): 사이즈 변수들을 세밀하게 조정했습니다
