+++ src/pages/MarketPage/AllProducts.module.scss
@@ -1,0 +1,7 @@
+.all-products {
+  width: 100%;
+  max-width: 1200px;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 20px;
+}
