+++ src/pages/MarketPage/BestProducts.module.scss
@@ -1,9 +1,3 @@
-.market-page {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-}
 .best-products {
   width: 100%;
   max-width: 1200px;
@@ -11,4 +5,11 @@
   display: flex;
   flex-wrap: wrap;
   gap: 20px;
+
+  &__titie {
+    width: 100%;
+    font-size: var(--font-title);
+    font-weight: bold;
+    margin-bottom: 20px;
+  }
 }
