+++ src/pages/MarketPage/MarketPage.module.scss
@@ -1,0 +1,13 @@
+.market-page {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  max-width: var(--market-container-max-width);
+  padding: var(--spacing-xs);
+  gap: var(--spacing-md);
+
+  @include tablet {
+    padding: var(--spacing-md);
+    gap: var(--spacing-lg);
+  }
+}
