+++ src/pages/MarketPage/PageNations.tsx
@@ -1,0 +1,4 @@
+function PageNations() {
+  return <div>{/* 페이지네이션 UI */}</div>;
+}
+export default PageNations;
