+++ src/styles/App.module.scss
@@ -5,5 +5,5 @@
 }
 .main {
   flex: 1;
-  padding: var(--spacing-md);
+  margin-top: 70px; /* Header height */
 }
