+++ src/styles/reset.scss
@@ -34,3 +34,11 @@ svg {
   display: block;
   max-width: 100%;
 }
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  margin: 0;
+}
