+++ src/styles/variables.scss
@@ -2,12 +2,23 @@
 :root {
   /* Font */
   --font-family-base: 'Pretendard', sans-serif;
-  --font-title: 2.5rem;
-  --font-title-sm: 2rem;
-  --font-title-xs: 1.5rem;
-  --font-content: 1.5rem;
-  --font-content-sm: 1rem;
-  --font-label: 1.125rem;
+  --font-title-lx: 2.5rem; // 40px
+  --font-title-lg: 2rem; // 32px
+  --font-title-md: 1.5rem; // 24px
+  --font-title-sm: 1.25rem; // 20px
+  --font-title-xs: 1.125rem; // 18px
+
+  --font-content-xl: 1.5rem; // 24px
+  --font-content-lg: 1.25rem; // 20px
+  --font-content-md: 1rem; // 16px
+  --font-content-sm: 0.875rem; // 14px
+  --font-content-xs: 0.75rem; // 12px
+
+  --font-label-xl: 1.25rem; // 20px
+  --font-label-lg: 1.125rem; // 18px
+  --font-label-md: 1rem; // 16px
+  --font-label-sm: 0.875rem; // 14px
+  --font-label-xs: 0.75rem; // 12px
 
   /* Color - Primary */
   --color-primary-100: #3692ff;
@@ -33,11 +44,11 @@
   --color-error: red;
 
   /* Spacing */
-  --spacing-xs: 1rem;
-  --spacing-sm: 1.125rem;
-  --spacing-md: 1.5rem;
-  --spacing-lg: 2.5rem;
-  --spacing-xl: 8rem;
+  --spacing-xs: 1rem; // 16px
+  --spacing-sm: 1.125rem; // 18px
+  --spacing-md: 1.5rem; // 24px
+  --spacing-lg: 2.5rem; // 40px
+  --spacing-xl: 8rem; // 128px
 
   /* Layout */
   --container-max-width: 1920px;
@@ -47,6 +58,7 @@
   --container-top-margin-sm: 80px;
   --auth-container-max-width: 640px;
   --auth-container-max-width-sm: 400px;
+  --market-container-max-width: 1200px;
   --tablet-max-width: 1199px;
   --mobile-max-width: 767px;
